+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - Update Craft CMS version 4.15.4 to 5.7.7
+- Add Ckeditor 5 default configs
+- Add module Tj Custom Admin Notices
+- Add module Tj Ckeditor Customizations
 
 ## [1.6.2.1] - 2025-05-20